Add Low Level IT in CassandraToSourceDbIT
