--- v2/spanner-to-sourcedb/src/test/resources/SpannerToCassandraSourceDbWideRowMaxColumnNameWithMaxTableNameIT/cassandra-max-col-schema.sql
@@ -1,4 +1,0 @@
-CREATE TABLE testtable_03tpcovf16ed0klxm3v808ch3btgq0uk (
-    id TEXT PRIMARY KEY,
-    col_qcbf69rmxtre3b_03tpcovf16ed TEXT
-);
