+++ v2/spanner-to-sourcedb/src/test/resources/SpannerToCassandraSourceIT/cassandra-schema.sql
@@ -216,3 +216,9 @@ CREATE TABLE EmptyStringJsonTable (
     duration_list_col LIST<TEXT>,
     frozen_ascii_set_col frozen<SET<ASCII>>
 );
+
+CREATE TABLE testtable_03tpcovf16ed0klxm3v808ch3btgq0uk (
+    id TEXT PRIMARY KEY,
+    col_qcbf69rmxtre3b_03tpcovf16ed TEXT
+);
+
